# Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EVADIA API",
        "description": "API complète de la plateforme de gestion hôtelière EVADIA. Comprend l'authentification, la gestion admin, le back-office hôtel et les endpoints client.",
        "contact": {
            "name": "EVADIA",
            "email": "contact@evadia.com"
        },
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/hotel/rooms": {
            "get": {
                "tags": [
                    "Hôtel - Chambres"
                ],
                "summary": "Liste des chambres",
                "description": "Retourne la liste des chambres/propriétés de l'hôtel avec photos, prix et statut actuels.",
                "operationId": "751e013c15818551ae00cf78ab8d4526",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filtrer par type de chambre",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des chambres",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nom": {
                                                        "type": "string",
                                                        "example": "Suite Deluxe"
                                                    },
                                                    "type_propriete": {
                                                        "type": "string",
                                                        "example": "suite"
                                                    },
                                                    "capacite": {
                                                        "type": "integer",
                                                        "example": 4
                                                    },
                                                    "nb_chambres": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "nb_lits": {
                                                        "type": "integer",
                                                        "example": 3
                                                    },
                                                    "superficie": {
                                                        "type": "integer",
                                                        "example": 55
                                                    },
                                                    "prix_actuel": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 180,
                                                        "nullable": true
                                                    },
                                                    "statut": {
                                                        "type": "string",
                                                        "example": "disponible",
                                                        "nullable": true
                                                    },
                                                    "photo_principale": {
                                                        "type": "string",
                                                        "example": "https://s3.../photo.jpg",
                                                        "nullable": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Hôtel - Chambres"
                ],
                "summary": "Créer une chambre",
                "description": "Ajoute une nouvelle chambre/propriété à l'hôtel.",
                "operationId": "72c97a3f16042790eae3de288e4ecbfc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "type_propriete",
                                    "capacite"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Suite Deluxe 201"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Suite spacieuse avec vue mer"
                                    },
                                    "type_propriete": {
                                        "type": "string",
                                        "example": "suite"
                                    },
                                    "capacite": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "nb_chambres": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "nb_lits": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "nb_salles_bain": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "superficie": {
                                        "type": "integer",
                                        "example": 55
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Chambre créée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Chambre créée avec succès"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/hotel/rooms/{id}": {
            "get": {
                "tags": [
                    "Hôtel - Chambres"
                ],
                "summary": "Détails d'une chambre",
                "description": "Retourne les informations complètes d'une chambre avec photos, équipements, prix et disponibilités.",
                "operationId": "5a903e19e0fd21f6f169729d00cb9357",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de la chambre",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Chambre non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Hôtel - Chambres"
                ],
                "summary": "Modifier une chambre",
                "operationId": "df2ab25f0cf0ff7f56e3aec5f8f77667",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Suite Deluxe 201"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "type_propriete": {
                                        "type": "string"
                                    },
                                    "capacite": {
                                        "type": "integer"
                                    },
                                    "nb_chambres": {
                                        "type": "integer"
                                    },
                                    "nb_lits": {
                                        "type": "integer"
                                    },
                                    "nb_salles_bain": {
                                        "type": "integer"
                                    },
                                    "superficie": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Chambre mise à jour"
                    },
                    "404": {
                        "description": "Chambre non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Hôtel - Chambres"
                ],
                "summary": "Supprimer une chambre",
                "operationId": "8909c8bfc76303060ddcbbe76c7d630d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chambre supprimée"
                    },
                    "404": {
                        "description": "Chambre non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token obtenu via /api/auth/login ou /api/auth/register",
                "bearerFormat": "Sanctum Token",
                "scheme": "bearer"
            }
        }
    },
...
}